feat: cleanup default screen and move location
